requirements.txt
flask
flask-cors
gunicorn
requests
firebase-admin
python-docx
PyPDF2
